requirements.txt
streamlit>=1.33,<2
torch>=2.2,<3
torchvision>=0.17,<1
easyocr>=1.7,<2
sentence-transformers>=2.6,<4
scikit-learn>=1.4,<2
nltk>=3.8,<4
textblob>=0.18,<1
numpy>=1.26,<3
opencv-python-headless>=4.9,<5
pillow>=10,<12
